requirements.txt
apachelogs==0.6.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
duckdb==1.2.2
duckdb_engine==0.17.0
idna==3.10
maxminddb==2.6.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pyarrow==20.0.0
pydicti==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
